requirements.txt
Package                Version
---------------------- -------------------
absl-py                0.11.0
argon2-cffi            20.1.0
astunparse             1.6.3
async-generator        1.10
attrs                  21.2.0
backcall               0.2.0
bleach                 3.3.0
brotlipy               0.7.0
cachetools             4.2.0
certifi                2020.12.5
cffi                   1.14.4
chardet                4.0.0
conda                  4.10.1
conda-package-handling 1.7.3
cryptography           3.4.7
cycler                 0.10.0
decorator              5.0.9
defusedxml             0.7.1
entrypoints            0.3
gast                   0.3.3
google-auth            1.24.0
google-auth-oauthlib   0.4.2
google-pasta           0.2.0
grpcio                 1.34.0
h5py                   2.10.0
idna                   2.10
ipykernel              5.5.5
ipython                7.23.1
ipython-genutils       0.2.0
ipywidgets             7.6.3
jedi                   0.18.0
Jinja2                 3.0.1
joblib                 1.0.1
jsonschema             3.2.0
jupyter                1.0.0
jupyter-client         6.1.12
jupyter-console        6.4.0
jupyter-core           4.7.1
jupyterlab-pygments    0.1.2
jupyterlab-widgets     1.0.0
Keras-Preprocessing    1.1.2
kiwisolver             1.3.1
Markdown               3.3.3
MarkupSafe             2.0.1
matplotlib             3.4.2
matplotlib-inline      0.1.2
mistune                0.8.4
nbclient               0.5.3
nbconvert              6.0.7
nbformat               5.1.3
nest-asyncio           1.5.1
notebook               6.4.0
numpy                  1.20.2
oauthlib               3.1.0
opt-einsum             3.3.0
packaging              20.9
pandas                 1.2.4
pandocfilters          1.4.3
parso                  0.8.2
pexpect                4.8.0
pickleshare            0.7.5
Pillow                 8.2.0
pip                    21.1.1
prometheus-client      0.10.1
prompt-toolkit         3.0.18
protobuf               3.14.0
ptyprocess             0.7.0
pyasn1                 0.4.8
pyasn1-modules         0.2.8
pycosat                0.6.3
pycparser              2.20
Pygments               2.9.0
pyOpenSSL              20.0.1
pyparsing              2.4.7
pyrsistent             0.17.3
PySocks                1.7.1
python-dateutil        2.8.1
pytz                   2021.1
pyzmq                  22.0.3
qtconsole              5.1.0
QtPy                   1.9.0
requests               2.25.1
requests-oauthlib      1.3.0
rsa                    4.6
ruamel-yaml-conda      0.15.80
scikit-learn           0.24.2
scipy                  1.4.1
Send2Trash             1.5.0
setuptools             49.6.0.post20210108
six                    1.16.0
sklearn                0.0
tensorboard            2.2.2
tensorboard-plugin-wit 1.7.0
tensorflow             2.2.0
tensorflow-estimator   2.2.0
termcolor              1.1.0
terminado              0.10.0
testpath               0.5.0
threadpoolctl          2.1.0
tornado                6.1
tqdm                   4.60.0
traitlets              5.0.5
urllib3                1.26.4
wcwidth                0.2.5
webencodings           0.5.1
Werkzeug               1.0.1
wheel                  0.36.2
widgetsnbextension     3.5.1
wrapt                  1.12.1
